# backend-contract/openapi.yaml
openapi: 3.0.3
info:
  title: AI Response API
  version: 0.4.0
servers:
  - url: http://127.0.0.1:8080
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is up
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Auth session
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Auth session
  /projects:
    get:
      summary: List user projects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Project list
    post:
      summary: Create new project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '200':
          description: Project created
  /projects/{projectId}/sources/text:
    post:
      summary: Upload and analyze text source
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, text]
              properties:
                title: { type: string }
                text: { type: string }
      responses:
        '200':
          description: Source analyzed
  /projects/{projectId}/sources/file:
    post:
      summary: Upload and analyze PDF/DOCX source
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File source analyzed
  /projects/{projectId}/context:
    get:
      summary: Get analyzed project context
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Project context summary and source analyses
  /ai/respond:
    post:
      summary: Stream answer based on selected project knowledge + live transcript
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, transcript]
              properties:
                projectId: { type: string }
                transcript: { type: string }
      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"delta":"Merhaba"}

                  data: {"delta":" dunya"}

                  data: {"done":true}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque